session-3-ai-agents/agents/tes-agent/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
chromadb>=0.4.0
